(engine) moved chart table and deathscreen in to the gamelib (ent_*).
