+++ src/ents_payer_death.h
@@ -1,0 +1,18 @@
+/*
+ * ents_payer_death.h
+ *
+ *  Created on: 9 авг. 2018 г.
+ *      Author: mastersan
+ */
+
+#ifndef SRC_ENTS_PAYER_DEATH_H_
+#define SRC_ENTS_PAYER_DEATH_H_
+
+void menu_death_on_enter(void);
+void menu_death_draw_on_enter(void);
+
+bool menu_death_on_event(int key);
+void menu_death_draw_on_update(void);
+
+
+#endif /* SRC_ENTS_PAYER_DEATH_H_ */
